refactor code setDate in helper
