+++ src/openWeather.js
@@ -1,4 +1,5 @@
 // import unitWeather from './unitChange';
+import { displayWeather, displayForecast } from './domManipulation';
 
 const apiWeather = () => {
   const api = {
@@ -50,34 +51,34 @@ const apiWeather = () => {
     }
   };
 
-  const displayWeather = (weather, unit) => {
-    const firstMessage = document.querySelector('.first-message');
-    firstMessage.style.display = 'none';
+  // const displayWeather = (weather, unit) => {
+  //   const firstMessage = document.querySelector('.first-message');
+  //   firstMessage.style.display = 'none';
 
-    const cityName = document.querySelector('.city-name');
-    cityName.textContent = `${weather.name}, ${weather.sys.country}`;
+  //   const cityName = document.querySelector('.city-name');
+  //   cityName.textContent = `${weather.name}, ${weather.sys.country}`;
 
-    const cityDate = document.querySelector('.city-date');
-    cityDate.textContent = setDate().weatherDate;
+  //   const cityDate = document.querySelector('.city-date');
+  //   cityDate.textContent = setDate().weatherDate;
 
-    const temperature = document.querySelector('.temperature');
-    temperature.textContent = `${weather.main.temp.toFixed(0)}°${unit}`;
+  //   const temperature = document.querySelector('.temperature');
+  //   temperature.textContent = `${weather.main.temp.toFixed(0)}°${unit}`;
 
-    const description = document.querySelector('.description');
-    description.textContent = `${weather.weather[0].description}`;
+  //   const description = document.querySelector('.description');
+  //   description.textContent = `${weather.weather[0].description}`;
 
-    const wind = document.querySelector('.wind');
-    wind.textContent = `${weather.wind.speed} km/h`;
+  //   const wind = document.querySelector('.wind');
+  //   wind.textContent = `${weather.wind.speed} km/h`;
 
-    const humidity = document.querySelector('.humidity');
-    humidity.textContent = `${weather.main.humidity}%`;
+  //   const humidity = document.querySelector('.humidity');
+  //   humidity.textContent = `${weather.main.humidity}%`;
 
-    const feelLike = document.querySelector('.feel-like');
-    feelLike.textContent = `${weather.main.feels_like.toFixed(0)}°${unit}`;
+  //   const feelLike = document.querySelector('.feel-like');
+  //   feelLike.textContent = `${weather.main.feels_like.toFixed(0)}°${unit}`;
 
-    const minMax = document.querySelector('.min-max');
-    minMax.textContent = `${weather.main.temp_min.toFixed(0)}°${unit} / ${weather.main.temp_max.toFixed(0)}°${unit}`;
-  };
+  //   const minMax = document.querySelector('.min-max');
+  //   minMax.textContent = `${weather.main.temp_min.toFixed(0)}°${unit} / ${weather.main.temp_max.toFixed(0)}°${unit}`;
+  // };
 
   async function getWeather(query, unit) {
     try {
@@ -94,75 +95,44 @@ const apiWeather = () => {
     }
   }
 
-  const setDate = () => {
-    const currentDate = new Date();
+  // const setForecastCard = (forecast, i, dayDescrip, unit) => {
+  //   const forecastCont = document.querySelector('.forecast');
+  //   const forecastCard = document.createElement('div');
+  //   forecastCard.classList.add('forecast-card', 'my-5', 'p-3', 'mx-2');
+  //   forecastCont.appendChild(forecastCard);
 
-    const days = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
-    const dayDescrip = (idx) => days[currentDate.getDay() + idx];
+  //   const forecastDay = document.createElement('div');
+  //   forecastDay.classList.add('forecast-day');
+  //   forecastDay.textContent = dayDescrip;
+  //   forecastCard.appendChild(forecastDay);
 
-    const months = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
-    const monthDescrip = months[currentDate.getMonth()];
+  //   const iconForecast = document.createElement('img');
+  //   iconForecast.classList.add('icon-forecast');
+  //   iconForecast.src = `http://openweathermap.org/img/w/${forecast.list[i].weather[0].icon}.png`;
+  //   forecastCard.appendChild(iconForecast);
 
-    const year = currentDate.getFullYear();
-    let month = '';
-    let day = '';
-    if (currentDate.getMonth() < 10) {
-      month = `0${currentDate.getMonth() + 1}`;
-    } else {
-      month = currentDate.getMonth() + 1;
-    }
-    if (currentDate.getDate() < 10) {
-      day = `0${currentDate.getDate()}`;
-    } else {
-      day = currentDate.getDate() + 1;
-    }
+  //   const tempForecast = document.createElement('div');
+  //   tempForecast.classList.add('temp-forecast');
+  //   tempForecast.textContent = `${forecast.list[i].main.temp.toFixed(0)}°${unit}`;
+  //   forecastCard.appendChild(tempForecast);
 
-    const weatherDate = `${dayDescrip(0)} ${day}, ${monthDescrip} ${year}`;
-    const forecastDate = `${year}-${month}-${day} 09:00:00`;
-
-    return {
-      weatherDate, forecastDate, dayDescrip, monthDescrip,
-    };
-  };
-
-  const setForecastCard = (forecast, i, dayDescrip, unit) => {
-    const forecastCont = document.querySelector('.forecast');
-    const forecastCard = document.createElement('div');
-    forecastCard.classList.add('forecast-card', 'my-5', 'p-3', 'mx-2');
-    forecastCont.appendChild(forecastCard);
-
-    const forecastDay = document.createElement('div');
-    forecastDay.classList.add('forecast-day');
-    forecastDay.textContent = dayDescrip;
-    forecastCard.appendChild(forecastDay);
-
-    const iconForecast = document.createElement('img');
-    iconForecast.classList.add('icon-forecast');
-    iconForecast.src = `http://openweathermap.org/img/w/${forecast.list[i].weather[0].icon}.png`;
-    forecastCard.appendChild(iconForecast);
-
-    const tempForecast = document.createElement('div');
-    tempForecast.classList.add('temp-forecast');
-    tempForecast.textContent = `${forecast.list[i].main.temp.toFixed(0)}°${unit}`;
-    forecastCard.appendChild(tempForecast);
-
-    const minMaxForecast = document.createElement('div');
-    minMaxForecast.classList.add('.min-max-forecast');
-    minMaxForecast.textContent = `${forecast.list[i].main.temp_min.toFixed(0)}°${unit} / ${forecast.list[i].main.temp_max.toFixed(0)}°${unit}`;
-  };
+  //   const minMaxForecast = document.createElement('div');
+  //   minMaxForecast.classList.add('.min-max-forecast');
+  //   minMaxForecast.textContent = `${forecast.list[i].main.temp_min.toFixed(0)}°${unit} / ${forecast.list[i].main.temp_max.toFixed(0)}°${unit}`;
+  // };
 
 
-  const displayForecast = (forecast, unit) => {
-    for (let i = 0; i < forecast.list.length; i += 1) {
-      if (forecast.list[i].dt_txt === setDate().forecastDate) {
-        setForecastCard(forecast, i, setDate().dayDescrip(1), unit);
-        setForecastCard(forecast, i + 8, setDate().dayDescrip(2), unit);
-        setForecastCard(forecast, i + 16, setDate().dayDescrip(3), unit);
-        setForecastCard(forecast, i + 24, setDate().dayDescrip(4), unit);
-        setForecastCard(forecast, i + 32, setDate().dayDescrip(5), unit);
-      }
-    }
-  };
+  // const displayForecast = (forecast, unit) => {
+  //   for (let i = 0; i < forecast.list.length; i += 1) {
+  //     if (forecast.list[i].dt_txt === setDate().forecastDate) {
+  //       setForecastCard(forecast, i, setDate().dayDescrip(1), unit);
+  //       setForecastCard(forecast, i + 8, setDate().dayDescrip(2), unit);
+  //       setForecastCard(forecast, i + 16, setDate().dayDescrip(3), unit);
+  //       setForecastCard(forecast, i + 24, setDate().dayDescrip(4), unit);
+  //       setForecastCard(forecast, i + 32, setDate().dayDescrip(5), unit);
+  //     }
+  //   }
+  // };
 
   async function getForecast(query, unit) {
     try {
